+++ data/auth/src/main/java/com/hilingual/data/auth/di/qualifier/TokenDataStore.kt
@@ -1,4 +1,4 @@
-package com.hilingual.data.auth.di.quilifier
+package com.hilingual.data.auth.di.qualifier
 
 import javax.inject.Qualifier
